fix: address sage review R2 on cortex#1185 (boundary + literal-true + DRY parser)
- ARCH: move review-flavors.ts src/bus → src/common/types (architecture.md §8 —
  config must not import src/bus). review-events re-exports it; the bus boundary
  is clean again (no src/common → src/bus import).
- SCHEMA: `review: z.literal(true).optional()` — `{prompt, review:false}` now
  fails at parse instead of slipping through as a prompt-only target (+ test).
- DRY: the `code-review.<flavor>` regex was still duplicated. Centralize as
  `reviewFlavorOfCapability` in review-flavors.ts; the schema and the bridge's
  reviewFlavorOf both delegate to it — one parser, one vocabulary.
- CHANGELOG: scope the routing claim to producer-side ("publishes on the subject
  the engine:sage consumer binds; consumer claim covered by the ReviewConsumer
  suite") instead of asserting an unproven end-to-end route.

tsc clean; carve-out clean; src/bus + config 1289 pass / 0 fail.
